Moved user2 signup function to its own file
