+++ functions/index.js
@@ -5,6 +5,7 @@ const sgMail = require('@sendgrid/mail');
 const caller = require('./tasks/caller.js');
 const fetch = require('./tasks/twilioFetch.js');
 const contactEmail = require('./tasks/contactEmail.js');
+const userTwo = require('./tasks/userTwo.js');
 const serviceAccount = require('./serviceAccountKey.json');
 
 sgMail.setApiKey(functions.config().sendgrid.key);
@@ -26,54 +27,8 @@ exports.callService = functions.pubsub
 
 exports.signupUserTwo = functions.firestore
   .document(`/users/{useruid}`)
-  .onCreate(async (snapshot, context) => {
-    const data = snapshot.data();
-    try {
-      await admin.auth().createUser({
-        email: data.email,
-        emailVerified: false,
-        password: Math.random()
-          .toString(36)
-          .slice(-8),
-        displayName: data.displayName,
-        photoURL: data.photoURL,
-        disabled: false,
-        uid: snapshot.id,
-      });
-
-      const passwordLink = await admin
-        .auth()
-        .generatePasswordResetLink(data.email);
-      console.log(passwordLink);
-      const msg = {
-        personalizations: [
-          {
-            to: [
-              {
-                email: data.email,
-                name: data.displayName,
-              },
-            ],
-            dynamic_template_data: {
-              url: passwordLink,
-            },
-            subject: 'A friend or loved one has signed you up for ReCaller!',
-          },
-        ],
-        from: {
-          email: 'labsrecaller@gmail.com',
-          name: 'ReCaller Team',
-        },
-        reply_to: {
-          email: 'labsrecaller@gmail.com',
-          name: 'ReCaller',
-        },
-        template_id: 'd-6077c121962b439f983a559b6f3a57f8',
-      };
-      sgMail.send(msg);
-    } catch (e) {
-      throw e;
-    }
+  .onCreate(snapshot => {
+    return userTwo.handler(admin, snapshot, sgMail);
   });
 
 exports.contactEmail = functions.firestore
